Stretch

+++ upd_script/update_all.sh
@@ -233,15 +233,15 @@ install_novnc() {
     feedback "--> Clone noVNC."
 
     if [ $VERSION -eq '8' ]; then
+        sudo rm -r noVNC
         sudo git clone  --depth=1 git://github.com/DexterInd/noVNC
         cd noVNC
-        sudo git pull
         sudo cp vnc_auto.html index.html
         #   # If we found Jesse, the proper location of the html files is in
         #   # /var/www/html
         sudo mkdir -p /var/www/html
         sudo cp -r $RASPBIAN_PATH/www /var/www/html
-        sudo mv -v /var/www/* /var/www/html/
+        sudo mv -f /var/www/* /var/www/html/
         sudo chmod +x /var/www/html/index.php
         sudo chmod +x /var/www/html/css/main.css
 
@@ -328,17 +328,23 @@ sudo sed -i "/dtparam=spi=on/d" /boot/config.txt
 sudo echo "dtparam=spi=on" >> /boot/config.txt
 sudo echo "dtparam=i2c_arm=on" >> /boot/config.txt
 
-# This is really imprtant for the BrickPi!
-sudo sed -i "/init_uart_clock=32000000/d" /boot/config.txt
-sudo echo "init_uart_clock=32000000" >> /boot/config.txt
 
-# Disable serial over UART
-sudo sed -i 's/console=ttyAMA0,115200//' /boot/cmdline.txt  #disable serial login on older images
-sudo sed -i 's/console=serial0,115200//' /boot/cmdline.txt  #disable serial login on the Pi3
-sudo sed -i 's/console=tty1//' /boot/cmdline.txt            #console=tty1 can also be there in the cmdline.txt file so remove that
-sudo sed -i 's/kgbdoc=ttyAMA0,115200//' /boot/cmdline.txt
-sudo systemctl stop serial-getty@ttyAMA0.service
-sudo systemctl disable serial-getty@ttyAMA0.service
+# Only reset UART on Jessie as we still support the BrickPi+
+if [ $VERSION -eq '8' ]
+then
+    # This is really imprtant for the BrickPi!
+    sudo sed -i "/init_uart_clock=32000000/d" /boot/config.txt
+    sudo echo "init_uart_clock=32000000" >> /boot/config.txt
+
+
+    # Disable serial over UART
+    sudo sed -i 's/console=ttyAMA0,115200//' /boot/cmdline.txt  #disable serial login on older images
+    sudo sed -i 's/console=serial0,115200//' /boot/cmdline.txt  #disable serial login on the Pi3
+    sudo sed -i 's/console=tty1//' /boot/cmdline.txt            #console=tty1 can also be there in the cmdline.txt file so remove that
+    sudo sed -i 's/kgbdoc=ttyAMA0,115200//' /boot/cmdline.txt
+    sudo systemctl stop serial-getty@ttyAMA0.service
+    sudo systemctl disable serial-getty@ttyAMA0.service
+fi
 
 feedback "--> End Kernel Updates."
 
@@ -425,8 +431,8 @@ install_novnc
 
 
 # feedback "Change bash permissions for desktop."
-# delete_line_from_file "xhost" /home/pi/.bashrc
-# add_line_to_end_of_file "xhost + >/dev/null" /home/pi/.bashrc
+delete_line_from_file "xhost" /home/pi/.bashrc
+add_line_to_end_of_file "xhost + >/dev/null 2>&1" /home/pi/.bashrc
 
 
 
@@ -434,7 +440,7 @@ install_novnc
 sleep 10
 
 ########################################################################
-## Last bit of house cleaning.
+## Last bit of house cleaning
 
 # Setup Hostname Changer
 feedback "--> Set up Hostname Changer."
@@ -531,7 +537,7 @@ if [ $VERSION -eq '9' ]
 then
     sudo apt-get install python-wxgtk3.0 python-psutil -y
 fi
-echo "Python-PSUtil"
+
 sudo apt-get clean -y		# Remove any unused packages.
 sudo apt-get autoremove -y 	# Remove unused packages.
 feedback "--> End cleanup."
